(related to bug 886) Floating point precision fixed.
git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@4292 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ experiments/statsound/src/java/client/pt/utl/ist/elab/client/statsound/displays/TableSoundVelocityModelProxy.java
@@ -75,7 +75,7 @@ public class TableSoundVelocityModelProxy extends com.linkare.rec.impl.client.ex
 			return String.valueOf(rowIndex + 1);
 		} else if (columnIndex == 1) {
 			// acquisition time
-			double time = rowIndex / 11.0;
+			double time = ((double) rowIndex) / ((double) 11.0);
 			return String.valueOf(Math.round(time));
 		}
 		PhysicsValue value = expDataModel.getValueAt(rowIndex, getColAtArray(columnIndex));
